+++ ElvUI_MerathilisUI/Core/General/Changelog.lua
@@ -12,9 +12,13 @@ local DISABLED_FONT_COLOR = DISABLED_FONT_COLOR
 
 local ChangeLogData = {
 	"Changes:",
-		"• Fix various compatibility errors with retail <=> classic",
-		"• Added back Rectangle Minimap (disabled by default)",
-		"• Move Tooltip Icons to Retail only due to incompatibility with Classic",
+		"• Fixed Pixel glow on RBReminder",
+		"• Added a reference spell to oUF_GCD",
+		"• Fixed Class Icon on the Character Skin",
+		"• Moved the Toggle Direction Option to the Embed Settings",
+		"• Fixed Enhanced NamePlate Auras",
+		"• Added a cool Evoker Class Crest. Thx Benik <3",
+		"• A lot of bag fixing",
 
 	" ",
 	"Notes:",